refactor: introduce useUploadCsv hook to consolidate upload logic
