+++ frontend/src/components/dashboard/create/sms/SMSRecipients.tsx
@@ -5,15 +5,8 @@ import React, {
   SetStateAction,
   useContext,
 } from 'react'
-import { useParams } from 'react-router-dom'
 import { OutboundLink } from 'react-ga'
 
-import {
-  uploadFileToS3,
-  deleteCsvStatus,
-  getCsvStatus,
-  CsvStatusResponse,
-} from 'services/upload.service'
 import {
   FileInput,
   CsvUpload,
@@ -31,9 +24,8 @@ import {
 import { LINKS } from 'config'
 import { i18n } from '@lingui/core'
 import { SMSCampaign, SMSPreview, SMSProgress } from 'classes'
-import { sendTiming } from 'services/ga.service'
 import { CampaignContext } from 'contexts/campaign.context'
-import useIsMounted from 'components/custom-hooks/use-is-mounted'
+import useUploadCsv from 'components/custom-hooks/use-upload-csv'
 
 import styles from '../Create.module.scss'
 
@@ -43,106 +35,30 @@ const SMSRecipients = ({
   setActiveStep: Dispatch<SetStateAction<SMSProgress>>
 }) => {
   const { campaign, updateCampaign } = useContext(CampaignContext)
-  const {
-    isCsvProcessing: initialIsProcessing,
-    numRecipients: initialNumRecipients,
-    csvFilename: initialCsvFilename,
-    demoMessageLimit,
-    params,
-  } = campaign as SMSCampaign
+  const { demoMessageLimit, params } = campaign as SMSCampaign
   const isDemo = !!demoMessageLimit
 
-  const [errorMessage, setErrorMessage] = useState(null)
-  const [isCsvProcessing, setIsCsvProcessing] = useState(initialIsProcessing)
-  const [isUploading, setIsUploading] = useState(false)
-  const [csvInfo, setCsvInfo] = useState<
-    Omit<CsvStatusResponse, 'isCsvProcessing' | 'preview'>
-  >({
-    numRecipients: initialNumRecipients,
-    csvFilename: initialCsvFilename,
-  })
-  const [preview, setPreview] = useState({} as SMSPreview)
-  const { id: campaignId } = useParams()
+  const [sampleCsvError, setSampleCsvError] = useState(null)
 
+  const {
+    isProcessing,
+    isUploading,
+    error,
+    preview,
+    csvInfo,
+    uploadFile,
+    clearCsvStatus,
+  } = useUploadCsv<SMSPreview>()
   const { csvFilename, numRecipients = 0 } = csvInfo
-  const isMounted = useIsMounted()
-
-  // Poll csv status
-  useEffect(() => {
-    if (!campaignId) return
-
-    let timeoutId: NodeJS.Timeout
-    const pollStatus = async () => {
-      try {
-        const { isCsvProcessing, preview, ...newCsvInfo } = await getCsvStatus(
-          +campaignId
-        )
-        // Prevent setting state if unmounted
-        if (!isMounted.current) return
-
-        setIsCsvProcessing(isCsvProcessing)
-        setCsvInfo(newCsvInfo)
-        if (preview) {
-          setPreview(preview as SMSPreview)
-        }
-        if (isCsvProcessing) {
-          timeoutId = setTimeout(pollStatus, 2000)
-        }
-      } catch (e) {
-        setErrorMessage(e.message)
-      }
-    }
-
-    // Retrieve status regardless of isCsvProcessing to retrieve csvError if any
-    // If completed, it will only poll once
-    pollStatus()
-
-    return () => clearTimeout(timeoutId)
-  }, [campaignId, isCsvProcessing, isMounted])
 
   // If campaign properties change, bubble up to root campaign object
   useEffect(() => {
-    updateCampaign({ isCsvProcessing, csvFilename, numRecipients })
-  }, [isCsvProcessing, csvFilename, numRecipients, updateCampaign])
-
-  // Handle file upload
-  async function uploadFile(files: File[]) {
-    setIsUploading(true)
-    setErrorMessage(null)
-    const uploadTimeStart = performance.now()
-
-    try {
-      // user did not select a file
-      if (!files[0] || !campaignId) {
-        return
-      }
-      clearCsvStatus()
-      const tempCsvFilename = await uploadFileToS3(+campaignId, files[0])
-
-      const uploadTimeEnd = performance.now()
-      sendTiming('Contacts file', 'upload', uploadTimeEnd - uploadTimeStart)
-
-      // Prevent setting state if unmounted
-      if (!isMounted.current) {
-        return
-      }
-
-      setIsCsvProcessing(true)
-      setCsvInfo((info) => ({ ...info, tempCsvFilename }))
-    } catch (err) {
-      setErrorMessage(err.message)
-    } finally {
-      setIsUploading(false)
-    }
-  }
-
-  // Hide csv error from previous upload and delete from db
-  function clearCsvStatus() {
-    if (campaignId) {
-      setCsvInfo((info) => ({ ...info, csvError: undefined }))
-      deleteCsvStatus(+campaignId)
-    }
-  }
+    updateCampaign({
+      isCsvProcessing: isProcessing,
+      csvFilename,
+      numRecipients,
+    })
+  }, [isProcessing, csvFilename, numRecipients, updateCampaign])
 
   return (
     <>
@@ -175,7 +91,7 @@ const SMSRecipients = ({
         )}
 
         <CsvUpload
-          isCsvProcessing={isCsvProcessing}
+          isCsvProcessing={isProcessing}
           csvInfo={csvInfo}
           onErrorClose={clearCsvStatus}
         >
@@ -184,7 +100,7 @@ const SMSRecipients = ({
           <SampleCsv
             params={params}
             defaultRecipient="81234567"
-            setErrorMsg={setErrorMessage}
+            setErrorMsg={setSampleCsvError}
           />
         </CsvUpload>
         {isDemo && (
@@ -195,10 +111,10 @@ const SMSRecipients = ({
             </span>
           </InfoBlock>
         )}
-        <ErrorBlock>{errorMessage}</ErrorBlock>
+        <ErrorBlock>{error || sampleCsvError}</ErrorBlock>
       </StepSection>
 
-      {!isCsvProcessing && numRecipients > 0 && (
+      {!isProcessing && numRecipients > 0 && (
         <StepSection>
           <p className={styles.greyText}>Message preview</p>
           <PreviewBlock body={preview.body}></PreviewBlock>
@@ -211,7 +127,7 @@ const SMSRecipients = ({
           onClick={() => setActiveStep((s) => s + 1)}
         />
         <TextButton
-          disabled={isCsvProcessing}
+          disabled={isProcessing}
           onClick={() => setActiveStep((s) => s - 1)}
         >
           Previous
