+++ frontend/src/components/dashboard/create/telegram/TelegramRecipients.tsx
@@ -5,16 +5,9 @@ import React, {
   Dispatch,
   SetStateAction,
 } from 'react'
-import { useParams } from 'react-router-dom'
 import { OutboundLink } from 'react-ga'
 
 import { CampaignContext } from 'contexts/campaign.context'
-import {
-  uploadFileToS3,
-  deleteCsvStatus,
-  getCsvStatus,
-  CsvStatusResponse,
-} from 'services/upload.service'
 import {
   FileInput,
   CsvUpload,
@@ -32,8 +25,7 @@ import {
 import { LINKS } from 'config'
 import { i18n } from '@lingui/core'
 import { TelegramPreview, TelegramProgress } from 'classes'
-import { sendTiming } from 'services/ga.service'
-import useIsMounted from 'components/custom-hooks/use-is-mounted'
+import useUploadCsv from 'components/custom-hooks/use-upload-csv'
 
 import styles from '../Create.module.scss'
 
@@ -43,103 +35,29 @@ const TelegramRecipients = ({
   setActiveStep: Dispatch<SetStateAction<TelegramProgress>>
 }) => {
   const { campaign, updateCampaign } = useContext(CampaignContext)
-  const {
-    isCsvProcessing: initialIsProcessing,
-    numRecipients: initialNumRecipients,
-    csvFilename: initialCsvFilename,
-    demoMessageLimit,
-    params,
-  } = campaign
+  const { demoMessageLimit, params } = campaign
   const isDemo = !!demoMessageLimit
+  const [sampleCsvError, setSampleCsvError] = useState(null)
 
-  const [errorMessage, setErrorMessage] = useState(null)
-  const [isCsvProcessing, setIsCsvProcessing] = useState(initialIsProcessing)
-  const [isUploading, setIsUploading] = useState(false)
-  const [csvInfo, setCsvInfo] = useState<
-    Omit<CsvStatusResponse, 'isCsvProcessing' | 'preview'>
-  >({
-    numRecipients: initialNumRecipients,
-    csvFilename: initialCsvFilename,
-  })
-  const [preview, setPreview] = useState({} as { body: string })
-  const { id: campaignId } = useParams()
-
+  const {
+    isProcessing,
+    isUploading,
+    error,
+    preview,
+    csvInfo,
+    uploadFile,
+    clearCsvStatus,
+  } = useUploadCsv<TelegramPreview>()
   const { csvFilename, numRecipients = 0 } = csvInfo
-  const isMounted = useIsMounted()
-
-  // Poll csv status
-  useEffect(() => {
-    if (!campaignId) return
-
-    let timeoutId: NodeJS.Timeout
-    const pollStatus = async () => {
-      try {
-        const { isCsvProcessing, preview, ...newCsvInfo } = await getCsvStatus(
-          +campaignId
-        )
-        // Prevent setting state if unmounted
-        if (!isMounted.current) return
-
-        setIsCsvProcessing(isCsvProcessing)
-        setCsvInfo(newCsvInfo)
-        if (preview) {
-          setPreview(preview as TelegramPreview)
-        }
-        if (isCsvProcessing) {
-          timeoutId = setTimeout(pollStatus, 2000)
-        }
-      } catch (e) {
-        setErrorMessage(e.message)
-      }
-    }
-
-    // Retrieve status regardless of isCsvProcessing to retrieve csvError if any
-    // If completed, it will only poll once
-    pollStatus()
-
-    return () => clearTimeout(timeoutId)
-  }, [campaignId, isCsvProcessing, isMounted])
 
   // If campaign properties change, bubble up to root campaign object
   useEffect(() => {
-    updateCampaign({ isCsvProcessing, csvFilename, numRecipients })
-  }, [isCsvProcessing, csvFilename, numRecipients, updateCampaign])
-
-  async function uploadFile(files: File[]) {
-    setIsUploading(true)
-    setErrorMessage(null)
-    const uploadTimeStart = performance.now()
-
-    try {
-      // user did not select a file
-      if (!files[0] || !campaignId) {
-        return
-      }
-      clearCsvStatus()
-      const tempCsvFilename = await uploadFileToS3(+campaignId, files[0])
-
-      const uploadTimeEnd = performance.now()
-      sendTiming('Contacts file', 'upload', uploadTimeEnd - uploadTimeStart)
-
-      // Prevent setting state if unmounted
-      if (!isMounted.current) return
-
-      setIsCsvProcessing(true)
-      setCsvInfo((info) => ({ ...info, tempCsvFilename }))
-    } catch (err) {
-      setErrorMessage(err.message)
-    } finally {
-      setIsUploading(false)
-    }
-  }
-
-  // Hide csv error from previous upload and delete from db
-  function clearCsvStatus() {
-    if (campaignId) {
-      setCsvInfo((info) => ({ ...info, csvError: undefined }))
-      deleteCsvStatus(+campaignId)
-    }
-  }
+    updateCampaign({
+      isCsvProcessing: isProcessing,
+      csvFilename,
+      numRecipients,
+    })
+  }, [isProcessing, csvFilename, numRecipients, updateCampaign])
 
   return (
     <>
@@ -172,7 +90,7 @@ const TelegramRecipients = ({
         )}
 
         <CsvUpload
-          isCsvProcessing={isCsvProcessing}
+          isCsvProcessing={isProcessing}
           csvInfo={csvInfo}
           onErrorClose={clearCsvStatus}
         >
@@ -181,7 +99,7 @@ const TelegramRecipients = ({
           <SampleCsv
             params={params}
             defaultRecipient="81234567"
-            setErrorMsg={setErrorMessage}
+            setErrorMsg={setSampleCsvError}
           />
         </CsvUpload>
 
@@ -193,10 +111,10 @@ const TelegramRecipients = ({
             </span>
           </InfoBlock>
         )}
-        <ErrorBlock>{errorMessage}</ErrorBlock>
+        <ErrorBlock>{error || sampleCsvError}</ErrorBlock>
       </StepSection>
 
-      {!isCsvProcessing && numRecipients > 0 && (
+      {!isProcessing && numRecipients > 0 && (
         <>
           <StepSection>
             <p className={styles.greyText}>Message preview</p>
@@ -209,11 +127,11 @@ const TelegramRecipients = ({
 
       <ButtonGroup>
         <NextButton
-          disabled={!numRecipients || isCsvProcessing}
+          disabled={!numRecipients || isProcessing}
           onClick={() => setActiveStep((s) => s + 1)}
         />
         <TextButton
-          disabled={isCsvProcessing}
+          disabled={isProcessing}
           onClick={() => setActiveStep((s) => s - 1)}
         >
           Previous
